# src/main/resources/static/api-doc.yml
openapi: '3.0.3'
info:
  title: Involvement System
  version: 1.0.0
servers:
  - url: http://localhost:8091
...
paths:
...
  /api/v1/users:
    get:
      summary: Получить список пользователей
      operationId: getUsers
      tags:
        - user
      parameters:
        - in: query
          name: page
          description: Номер страницы (нумерация начинается с 1)
          allowEmptyValue: true
          schema:
            type: integer
            default: 1
          example: 1
        - in: query
          name: size
          description: Размер страницы (сколько пользователей выведется на страницу)
          allowEmptyValue: true
          schema:
            type: integer
            default: 20
          example: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
    post:
      summary: Создать нового пользователя
      operationId: createUser
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/v1/users/{id}:
    get:
      summary: Получить пользователя по идентификатору
      operationId: getUserById
      tags:
        - user
      parameters:
        - in: path
          name: id
          description: Идентификатор пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 333
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    put:
      summary: Изменить пользователя
      operationId: updateUser
      tags:
        - user
      parameters:
        - in: path
          name: id
          description: Идентификатор пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 333
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      summary: Удалить пользователя
      operationId: deleteUser
      tags:
        - user
      parameters:
        - in: path
          name: id
          description: Идентификатор пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 333
      responses:
        '200':
          description: OK
  /api/v1/users/{id}/enable:
    patch:
      summary: Активизировать или отключить пользователя
      operationId: enableUser
      tags:
        - user
      parameters:
        - in: path
          name: id
          description: Идентификатор пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 333
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEnableRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
components:
  schemas:
...
    UserRequest:
      type: object
      description: Запрос на создание пользователя
      properties:
        fullName:
          type: string
          description: Полное имя пользователя
          example: Иванов Иван Иванович
        login:
          type: string
          description: Логин пользователя
          example: ivaniv_ii
        phone:
          type: string
          description: Телефон пользователя
          example: +7(999)0010001
        email:
          type: string
          description: Электронный адрес пользователя
          example: ivaniv_ii@our-company.mock
        departmentId:
          type: integer
          format: int64
          description: Идентификатор отдела
          example: 6
        positionId:
          type: integer
          format: int64
          description: Идентификатор должности
          example: 6
        roles:
          type: array
          items:
            type: string
          description: Роли. Доступные роли ADMIN, MANAGER, HR_MANAGER, USER
          example: [MANAGER, USER]
    UserEnableRequest:
      type: object
      description: Запрос на активизацию или отключение пользователя
      properties:
        enable:
          type: boolean
          description: true - активизировать пользователя, false - отключить пользователя
          example: false
    UserResponse:
      type: object
      description: Пользователь
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор задания
          example: 45
        fullName:
          type: string
          description: Полное имя пользователя
          example: Иванов Иван Иванович
        login:
          type: string
          description: Логин пользователя
          example: ivaniv_ii
        phone:
          type: string
          description: Телефон пользователя
          example: +7(999)0010001
        email:
          type: string
          description: Электронный адрес пользователя
          example: ivaniv_ii@our-company.mock
        department:
          type: string
          description: Наименование отдела
          example: Отдел разработки
        position:
          type: string
          description: Наименование должности
          example: Ведущий разработчик
        roles:
          type: array
          items:
            type: string
          description: Роли.
          example: [MANAGER, USER]
    UserPageResponse:
      type: object
      description: Страница списка пользователей
      properties:
        content:
          description: Список пользователей
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          description: Номер страницы
          type: integer
          example: 1
        size:
          description: Размер страницы (количество пользователей на странице)
          type: integer
          example: 20
        totalItems:
          description: Всего пользователей
          type: integer
          format: int64
          example: 2
        totalPages:
          description: Всего страниц
          type: integer
          example: 1